+++ database/models_webhook.py
@@ -2,7 +2,7 @@
 
 from datetime import datetime
 
-from sqlalchemy import Boolean, Column, DateTime, Integer, String, Text
+from sqlalchemy import BigInteger, Boolean, Column, DateTime, Integer, String, Text
 
 from database.db import Base
 
@@ -12,7 +12,7 @@ class WebhookProcessingLog(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     created_at = Column(DateTime, default=datetime.utcnow, index=True)
-    user_id = Column(Integer, nullable=True, index=True)
+    user_id = Column(BigInteger, nullable=True, index=True)  # stores telegram_id
     event_type = Column(String(120), nullable=True)
     success = Column(Boolean, nullable=False, default=False)
     detail = Column(Text, nullable=True)
